Move auth controllers from user module to the authentication module
