+++ modules/authentication/server-scala/src/main/scala/routes/FacebookAuthController.scala
@@ -1,4 +1,4 @@
-package routes.auth
+package routes
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.model.headers.HttpCookie
@@ -11,46 +11,44 @@ import common.errors.AmbigousResult
 import common.implicits.RichDBIO._
 import common.implicits.RichFuture._
 import model.User
-import model.oauth.google.{GoogleAuth, GoogleOauth2Response}
+import model.facebook.{FacebookAuth, FacebookOauth2Response}
 import modules.jwt.model.JwtContent
-import modules.jwt.model.Tokens._
 import modules.jwt.service.JwtAuthService
 import org.mindrot.jbcrypt.BCrypt
-import repositories.UserRepository
-import repositories.auth.GoogleAuthRepository
+import repositories.{FacebookAuthRepository, UserRepository}
 import services.ExternalApiService
 import spray.json._
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
 
-class GoogleAuthController @Inject()(@Named("google") oauth2Service: OAuth20Service,
-                                     externalApiService: ExternalApiService,
-                                     userRepository: UserRepository,
-                                     googleAuthRepository: GoogleAuthRepository,
-                                     jwtAuthService: JwtAuthService[JwtContent])
-                                    (implicit val executionContext: ExecutionContext) {
+class FacebookAuthController @Inject()(@Named("facebook") oauth2Service: OAuth20Service,
+                                       externalApiService: ExternalApiService,
+                                       userRepository: UserRepository,
+                                       facebookAuthRepository: FacebookAuthRepository,
+                                       jwtAuthService: JwtAuthService[JwtContent])
+                                      (implicit val executionContext: ExecutionContext) {
 
   val routes: Route =
-    (path("auth" / "google") & get) {
+    (path("auth" / "facebook") & get) {
       redirect(oauth2Service.getAuthorizationUrl, StatusCodes.Found)
-    } ~ (path("auth" / "google" / "callback") & get) {
+    } ~ (path("auth" / "facebook" / "callback") & get) {
       parameters('state.?, 'code) { (state, code) =>
         onComplete {
           for {
-            googleAuthInfo <- externalApiService.getUserInfo[GoogleOauth2Response](code, "https://content.googleapis.com/oauth2/v2/userinfo", oauth2Service)
-            user <- googleAuthRepository.findOne(googleAuthInfo.id).run.flatMap {
-              case Some(gUser) => userRepository.findOne(gUser.userId).run failOnNone AmbigousResult(s"User with id: ${gUser.userId} stored in google auth table, but not in the user table")
+            facebookAuthInfo <- externalApiService.getUserInfo[FacebookOauth2Response](code, "https://graph.facebook.com/me?fields=id,name,email", oauth2Service)
+            user <- facebookAuthRepository.findOne(facebookAuthInfo.id).run.flatMap {
+              case Some(fUser) => userRepository.findOne(fUser.userId).run failOnNone AmbigousResult(s"User with id: ${fUser.userId} stored in facebook auth table, but not in the user table")
               case None => for {
                 user <- userRepository.save(
                   User(
-                    username = googleAuthInfo.name,
-                    email = googleAuthInfo.email,
-                    password = BCrypt.hashpw(googleAuthInfo.id, BCrypt.gensalt),
+                    username = facebookAuthInfo.name,
+                    email = facebookAuthInfo.email,
+                    password = BCrypt.hashpw(facebookAuthInfo.id, BCrypt.gensalt),
                     role = "user",
                     isActive = true)
                 ).run
-                _ <- googleAuthRepository.save(GoogleAuth(Some(googleAuthInfo.id), googleAuthInfo.name, user.id.get)).run
+                _ <- facebookAuthRepository.save(FacebookAuth(Some(facebookAuthInfo.id), facebookAuthInfo.name, user.id.get)).run
               } yield user
             }
           } yield jwtAuthService.createTokens(JwtContent(user.id.get), user.password)
