+++ modules/authentication/server-scala/src/main/scala/routes/GoogleAuthController.scala
@@ -1,4 +1,4 @@
-package routes.auth
+package routes
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.model.headers.HttpCookie
@@ -11,46 +11,44 @@ import common.errors.AmbigousResult
 import common.implicits.RichDBIO._
 import common.implicits.RichFuture._
 import model.User
-import model.oauth.linkedin.{LinkedinAuth, LinkedinOauth2Response}
+import model.google.{GoogleAuth, GoogleOauth2Response}
 import modules.jwt.model.JwtContent
-import modules.jwt.model.Tokens._
 import modules.jwt.service.JwtAuthService
 import org.mindrot.jbcrypt.BCrypt
-import repositories.UserRepository
-import repositories.auth.LinkedinAuthRepository
+import repositories.{GoogleAuthRepository, UserRepository}
 import services.ExternalApiService
 import spray.json._
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
 
-class LinkedinAuthController @Inject()(@Named("linkedin") oauth2Service: OAuth20Service,
-                                       externalApiService: ExternalApiService,
-                                       userRepository: UserRepository,
-                                       linkedinAuthRepository: LinkedinAuthRepository,
-                                       jwtAuthService: JwtAuthService[JwtContent])
-                                      (implicit val executionContext: ExecutionContext) {
+class GoogleAuthController @Inject()(@Named("google") oauth2Service: OAuth20Service,
+                                     externalApiService: ExternalApiService,
+                                     userRepository: UserRepository,
+                                     googleAuthRepository: GoogleAuthRepository,
+                                     jwtAuthService: JwtAuthService[JwtContent])
+                                    (implicit val executionContext: ExecutionContext) {
 
   val routes: Route =
-    (path("auth" / "linkedin") & get) {
+    (path("auth" / "google") & get) {
       redirect(oauth2Service.getAuthorizationUrl, StatusCodes.Found)
-    } ~ (path("auth" / "linkedin" / "callback") & get) {
+    } ~ (path("auth" / "google" / "callback") & get) {
       parameters('state.?, 'code) { (state, code) =>
         onComplete {
           for {
-            linkedinAuthInfo <- externalApiService.getUserInfo[LinkedinOauth2Response](code, "https://api.linkedin.com/v2/me", oauth2Service)
-            user <- linkedinAuthRepository.findOne(linkedinAuthInfo.id).run.flatMap {
-              case Some(lnUser) => userRepository.findOne(lnUser.userId).run failOnNone AmbigousResult(s"User with id: ${lnUser.userId} stored in linkedin auth table, but not in the user table")
+            googleAuthInfo <- externalApiService.getUserInfo[GoogleOauth2Response](code, "https://content.googleapis.com/oauth2/v2/userinfo", oauth2Service)
+            user <- googleAuthRepository.findOne(googleAuthInfo.id).run.flatMap {
+              case Some(gUser) => userRepository.findOne(gUser.userId).run failOnNone AmbigousResult(s"User with id: ${gUser.userId} stored in google auth table, but not in the user table")
               case None => for {
                 user <- userRepository.save(
                   User(
-                    username = linkedinAuthInfo.name,
-                    email = linkedinAuthInfo.email,
-                    password = BCrypt.hashpw(linkedinAuthInfo.id, BCrypt.gensalt),
+                    username = googleAuthInfo.name,
+                    email = googleAuthInfo.email,
+                    password = BCrypt.hashpw(googleAuthInfo.id, BCrypt.gensalt),
                     role = "user",
                     isActive = true)
                 ).run
-                _ <- linkedinAuthRepository.save(LinkedinAuth(Some(linkedinAuthInfo.id), linkedinAuthInfo.name, user.id.get)).run
+                _ <- googleAuthRepository.save(GoogleAuth(Some(googleAuthInfo.id), googleAuthInfo.name, user.id.get)).run
               } yield user
             }
           } yield jwtAuthService.createTokens(JwtContent(user.id.get), user.password)
