+++ modules/authentication/server-scala/src/main/scala/routes/LinkedinAuthController.scala
@@ -1,4 +1,4 @@
-package routes.auth
+package routes
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.model.headers.HttpCookie
@@ -11,46 +11,44 @@ import common.errors.AmbigousResult
 import common.implicits.RichDBIO._
 import common.implicits.RichFuture._
 import model.User
-import model.oauth.github.{GithubAuth, GithubOauth2Response}
+import model.linkedin.{LinkedinAuth, LinkedinOauth2Response}
 import modules.jwt.model.JwtContent
-import modules.jwt.model.Tokens._
 import modules.jwt.service.JwtAuthService
 import org.mindrot.jbcrypt.BCrypt
-import repositories.UserRepository
-import repositories.auth.GithubAuthRepository
+import repositories.{LinkedinAuthRepository, UserRepository}
 import services.ExternalApiService
 import spray.json._
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
 
-class GithubAuthController @Inject()(@Named("github") oauth2Service: OAuth20Service,
-                                     externalApiService: ExternalApiService,
-                                     userRepository: UserRepository,
-                                     githubAuthRepository: GithubAuthRepository,
-                                     jwtAuthService: JwtAuthService[JwtContent])
-                                    (implicit val executionContext: ExecutionContext) {
+class LinkedinAuthController @Inject()(@Named("linkedin") oauth2Service: OAuth20Service,
+                                       externalApiService: ExternalApiService,
+                                       userRepository: UserRepository,
+                                       linkedinAuthRepository: LinkedinAuthRepository,
+                                       jwtAuthService: JwtAuthService[JwtContent])
+                                      (implicit val executionContext: ExecutionContext) {
 
   val routes: Route =
-    (path("auth" / "github") & get) {
+    (path("auth" / "linkedin") & get) {
       redirect(oauth2Service.getAuthorizationUrl, StatusCodes.Found)
-    } ~ (path("auth" / "github" / "callback") & get) {
+    } ~ (path("auth" / "linkedin" / "callback") & get) {
       parameters('state.?, 'code) { (state, code) =>
         onComplete {
           for {
-            githubAuthInfo <- externalApiService.getUserInfo[GithubOauth2Response](code, "https://api.github.com/user", oauth2Service)
-            user <- githubAuthRepository.findOne(githubAuthInfo.id).run.flatMap {
-              case Some(ghUser) => userRepository.findOne(ghUser.userId).run failOnNone AmbigousResult(s"User with id: ${ghUser.userId} stored in github auth table, but not in the user table")
+            linkedinAuthInfo <- externalApiService.getUserInfo[LinkedinOauth2Response](code, "https://api.linkedin.com/v2/me", oauth2Service)
+            user <- linkedinAuthRepository.findOne(linkedinAuthInfo.id).run.flatMap {
+              case Some(lnUser) => userRepository.findOne(lnUser.userId).run failOnNone AmbigousResult(s"User with id: ${lnUser.userId} stored in linkedin auth table, but not in the user table")
               case None => for {
                 user <- userRepository.save(
                   User(
-                    username = githubAuthInfo.name,
-                    email = githubAuthInfo.email,
-                    password = BCrypt.hashpw(githubAuthInfo.id.toString + githubAuthInfo.name, BCrypt.gensalt),
+                    username = linkedinAuthInfo.name,
+                    email = linkedinAuthInfo.email,
+                    password = BCrypt.hashpw(linkedinAuthInfo.id, BCrypt.gensalt),
                     role = "user",
                     isActive = true)
                 ).run
-                _ <- githubAuthRepository.save(GithubAuth(Some(githubAuthInfo.id), githubAuthInfo.name, user.id.get)).run
+                _ <- linkedinAuthRepository.save(LinkedinAuth(Some(linkedinAuthInfo.id), linkedinAuthInfo.name, user.id.get)).run
               } yield user
             }
           } yield jwtAuthService.createTokens(JwtContent(user.id.get), user.password)
